+++ modules/iam-access-groups/src/main/java/com/ibm/cloud/platform_services/iam_access_groups/v2/model/Group.java
@@ -12,6 +12,8 @@
  */
 package com.ibm.cloud.platform_services.iam_access_groups.v2.model;
 
+import java.util.Date;
+
 import com.google.gson.annotations.SerializedName;
 import com.ibm.cloud.sdk.core.service.model.GenericModel;
 
@@ -26,11 +28,11 @@ public class Group extends GenericModel {
   @SerializedName("account_id")
   protected String accountId;
   @SerializedName("created_at")
-  protected String createdAt;
+  protected Date createdAt;
   @SerializedName("created_by_id")
   protected String createdById;
   @SerializedName("last_modified_at")
-  protected String lastModifiedAt;
+  protected Date lastModifiedAt;
   @SerializedName("last_modified_by_id")
   protected String lastModifiedById;
   protected String href;
@@ -88,7 +90,7 @@ public class Group extends GenericModel {
    *
    * @return the createdAt
    */
-  public String getCreatedAt() {
+  public Date getCreatedAt() {
     return createdAt;
   }
 
@@ -110,7 +112,7 @@ public class Group extends GenericModel {
    *
    * @return the lastModifiedAt
    */
-  public String getLastModifiedAt() {
+  public Date getLastModifiedAt() {
     return lastModifiedAt;
   }